task13

+++ src/com/company/Stage.java
@@ -2,7 +2,7 @@ package com.company;
 
 public abstract class Stage {
 
-    protected int length;
+    protected float length;
     protected String description;
     public String getDescription() {
         return description;
